+++ Magic/CHANGELOG.md
@@ -47,6 +47,7 @@
  - Re-enabled and re-made Force, hopefully less OP now but still fun and interesting, now available to Wizard rank
  - The Retarget action was re-written to allow it to change any targeting parameters, such as `target: block`
  - Most spell messages now go to the action bar, cooldown messages now shown for "quiet" wands (the default)
+ - The Damage action now bypasses vanilla damage cooldowns. Use `no_damage_ticks: -1` to turn this off.
 
 ## Additions